fix(recurring): keep a schedule alive through our own outages and stop sub-millisecond claims from stalling it
A relayer that is switched off or rejects our credentials is a platform
problem, so it now retries forever instead of counting toward the owner's
pause threshold, and the owner-facing copy says the schedule stays active.
The subscription claim compared next_charge_at against a round-tripped JS
Date, which truncates to the millisecond and silently never matched a row
with sub-millisecond precision; it now re-tests dueness server-side.

+++ api/dca-strategies.js
@@ -305,7 +305,7 @@ export default wrap(async (req, res) => {
 					LIMIT 1
 				) AS last_execution,
 				(
-					SELECT COUNT(*) FROM dca_executions e
+					SELECT COUNT(*)::int FROM dca_executions e
 					WHERE e.strategy_id = s.id AND e.status = 'success'
 				) AS executions_total
 			FROM dca_strategies s
